refactor:レンダリングを最適化 #19
