+++ src/components/room/Room.js
@@ -13,23 +13,7 @@ import { useNavigate } from "react-router-dom";
 // import "./css/Room.css";
 import RoomNavbar from "./RoomNavbar";
 import RoomDetail from "../common/RoomDetail";
-
-// 時間表示を○○：○○にする関数
-const formatHHMM = (time) => {
-  const date = new Date(time);
-  const hours = date.getHours();
-  const minutes = date.getMinutes();
-
-  return `${hours}:${minutes.toString().padStart(2, "0")}`;
-};
-
-const formatHHMMforTimeStamp = (timestamp) => {
-  const date = new Date(timestamp.seconds * 1000);
-  const hours = date.getHours();
-  const minutes = date.getMinutes();
-
-  return `${hours}:${minutes.toString().padStart(2, "0")}`;
-};
+import PostItem from "./PostItem";
 
 const Room = () => {
   const { roomID, userID } = useContext(Context);
@@ -155,80 +139,9 @@ const Room = () => {
       <button onClick={openRoomDetail}>{roomName}</button>
       <div className="postContainer">
         {/* ポストリストの各ポストごとに描画 */}
-        {postList.map((post) => {
-          // console.log("Current User:", currentUser.uid);
-          // console.log("Post Author:", post.author);
-          const authorUser = userList.find((user) => user.id === post.author);
-          const authorName = authorUser ? authorUser.username : "Unknown User";
-          const authorLevel = authorUser ? "Level " + authorUser.level : "";
-          const goalTime = post.goalTime
-            ? `${formatHHMMforTimeStamp(post.goalTime)}`
-            : "";
-
-          return (
-            // ポストの投稿者が自分かそれ以外かでクラスを変える
-            <div
-              className={`post_${userID === post.author ? "r" : "l"}`}
-              key={post.id}
-            >
-              <div className="post">
-                <div className="user">
-                  <div className="userIcon">
-                    <p>{authorName}</p>
-                  </div>
-                  <p className="userLevel">{authorLevel}</p>
-                </div>
-                {/* ポストの種類によってスタンプを変える */}
-                <div className="stamp">
-                  {post.type === "startBath" ? (
-                    <img
-                      src="/nyuyokutyu.png"
-                      alt="入浴"
-                      width="200px"
-                      height="200px"
-                    />
-                  ) : (
-                    ""
-                  )}
-                  {post.type === "cancelBath" ? (
-                    <img
-                      src="/cancel.png"
-                      alt="お風呂キャンセル"
-                      width="200px"
-                      height="200px"
-                    />
-                  ) : (
-                    ""
-                  )}
-                  {post.type === "endBath" ? (
-                    <img
-                      src="agatta.png"
-                      alt="上がった!"
-                      width="200px"
-                      height="200px"
-                    />
-                  ) : (
-                    ""
-                  )}
-                  {post.type === "setBathGoal" ? (
-                    <div>
-                      <h3 className="goalTime">{goalTime}</h3>
-                      <img
-                        src="setGoalTime.png"
-                        alt="にお風呂に入る!"
-                        width="200px"
-                        height="200px"
-                      />
-                    </div>
-                  ) : (
-                    ""
-                  )}
-                  <p className="timeStamp">{formatHHMM(post.date)}</p>
-                </div>
-              </div>
-            </div>
-          );
-        })}
+        {postList.map((post) => (
+          <PostItem post={post} userList={userList}/>
+        ))}
       </div>
       <RoomNavbar lastPostType={lastPostType} />
